Fix: cannot assign actions to IDs without animdata
This bug was introduced in #126561. The issue was that the UI action
selectors no longer ensured that animdata existed before attempting
to assign an action, which would then cause the assignment to fail.
However, this was only the case in Blender builds without experimental
features, since the new Baklava code paths would otherwise be taken,
which *do* handle this correctly.

This fixes the issue by ensuring animdata in the non-experimental code
paths again.