Fix: assigning action to an action constraint doesn't restrict its type
This is a followup to #128072, which fixed the main bug but left the
assigned action without its `idroot` set, so the same action could get
assigned to e.g. both an action constraint and lamp data if done in
just the right way.

This fixes this secondary issue by setting `idroot` on an action if not
already set when assigning it to an action constraint.