WIP: Optimization: Avoid allocating large link lists when iterating over instances for snapping
Issue observed when investigating #125646

In the file attached to that report, a large number of raycasts were
operated to identify edge occlusion.

The problem involved dealing with the instances in the file.

To minimize the problem, the solution proposed at the moment is not to
allocate large link lists, but to use a callback to iterate over
instances.