Vulkan: Log selected device
Currently the log only contained the first compatible device. It is
more important to the user to know which device is used.

This PR increases the level of the first compatible device so it is only
visible when increasing the log level. It reports the device, driver and
vendor when starting blender with `--debug-gpu`.