Anim: Show Action Slot type in panel + change type icon RNA wrapper name
This PR combines two changes, both necessary to properly show the Action
Slot ID type in the GUI:

1. Expose the slot's ID type as an RNA enum property `id_root`, in the
   same way that the DNA `Action::idtype` is exposed as `Action.id_root`
   in RNA. The only differences are that this property is read-only on
   slots (it is supposed to be correct, and if it's not, I'd prefer a
   bug report over people working around this), and that the default
   value is named differently (`ANY` on Actions, `UNSPECIFIED` on
   Slots). The latter is done to clarify that this is not a situation
   that will last for long; as soon as the slot is assigned to any ID,
   it will be locked to that ID type.

2. Rename the RNA property `ActionSlot.idtype_icon` to
   `ActionSlot.id_root_icon` to be consistent with the above.

And of course the type is drawn in the Slot properties panel in the
Action editor. This now replaces the icon in the slot name (in the same
panel), so that it's clearer that that is the slot name, and not an
object/material/whatevertype name.