Fix #128138: Stroke paint curves are missing in the list
a1b4d5ecc8 added these extra rule enforcements, but editable linked IDs
have since been added, and the rules haven't been re-evaluated for this
case.

I think the intention of this check was to disallow assigning an ID
pointer to a non-editable ID. In the past that meant a non-linked ID,
now there's a `ID_IS_EDITABLE()` check that is semantically more
correct, and accounds for the case of linked but editable assets.