Function Node: Compare Node, add epsilon socket to integer data type
Typically the idea of an epsilon is used to measure a small amount.
However, in practical use, this is used to compare a range of values.

This patch adds an Epsilon socket to the Integer mode.
All other data types except String have this socket.

The new socket is an Integer value with a default_value of zero.

This can be achieved with the Math node using the Compare function
but this patch adds support to the Compare node and the Equal and
Not Equal operations.

Example usage of Math node rather than dedicated Compare node.
https://youtu.be/-aJahfHZpzg?si=HolztkPir3LjVsZq&t=103