Geometry Nodes: UI: Add dynamic node class to switch nodes
This patch extends the existing use of node class to switches.
The aim is to improve node tree readability.
Dynamic changes based on data type already exist for the
Mix and Map Range nodes.

The scope of changes is limited to switches, random value and
some additional vector nodes.

Hash and compare nodes are out of scope because the inputs
and outputs are different types.

There is no plan to add additional node classes.