Fix #128247: Show warning when applying rotation on invalid matrix
When a object matrix has a determinant of 0 (which typically means 0s on
one or more of the scale axes), the matrix inversion would become
invalid thus the inversion matrix will be quietly set to diag(0). This
would basically reset all vertices to the origin of the object after
applying the inversion matrix. This happens in the "apply rotation"
stage.

Now addtional warnings will be shown to the user when there are
objects that tries to go through such operation, and they will be
skipped.