IO: Add file unit type selector to stl importer and exporter
As STL is a unitless format, most other programs assume it to be millimeters. Blender on the other hand uses meters. To allow for CAD modelling and exporting to STL in real life measures, this commit supplies the user with a dropdown allowing to choose the target unit type (m, dm, cm, mm, in), whilst leaving a custom option for user defined scale (otherwise scale factor input field is disabled); the chosen unit will modify the scale factor value.
The default unit stays meter (factor 1) to not introduce any breaking changes to existing workflows.