IO: add scene unit option to stl io unit type selctor
This commit adds a new option to the dropdown to set the scale factor accordingly to the used scene unit of the blend file (blender assumes meter, thus a conversion from meter to unit system is needed). This is useful if the user already decides to model in the target unit system used by the other application.
Also done some cleanup, removed some redundant code.