Function Node: Value to String Integer subtype
This patch adds an `Integer` type to the Value to String node.
This allows for converting integer inputs into string outputs without
any loss of precision. Something which the original implementation,
now labelled `Float`, is not able to do.

This issue becomes evident at values greater than `2 ^ 24` or
`16777216`, where the precision loss is high enough that the float
implementation starts skipping certain integers.