Fix #127093: Kuwahara node produces NaNs
The Kuwahara node produces NaNs if the pixel has a very high value and
sharpness is also high. This is because the weighted sum of the Kuwahara
sectors can have a zero total weight, which causes zero division. To fix
this, we clamp the weights to a small minimum number.