Fix uninitialized memory after realizing instance.
This is (arguably) a bug with the SpanAttributeWriter, which always
overwrites the entire destination varray with the buffer span,
regardless of which values were actually written. When moving the
physics world data, the moved body-, constraint-, and shape-attributes
are not overwritten, but the SpanAttributeWriter still replaces them
with uninitialized memory.

Now the `finish()` function of the writer supports a version with an
index mask.