Anim: fix Action slot idtype endianness switching
`ActionSlot::idtype` needs some extra care when reading from a blend
file. Blender's generic endian switching needs to be un-done, as the ID
type values are not numerically the same between little and big endian
machines. Due to the way they are defined, they are always in the same
byte order, regardless of hardware/platform endianness.

This is the same for Action Slots as #128129 does for Actions.