Nodes: Add Negate operator to Math and Vector Math nodes
This adds a 1-op negate function to the math nodes.

Whilst it can be constructed from a multiply node this has
the advantage of being cleaner from a UI perspective and
potentially more performant.

It aids node tree readability and is easier to debug as there
is only one socket available.