Fix #76374: Multires artifacts when sculping on lower level
In the process of propagating changes from a lower level to a higher
level, the process needs to avoid applying the existing detail to the
grid coordinates at the corners of the current subdivision level. The
delta between these points is entirely due to the brush stroke, so
reapplying this has the effect of undoing the displacement.

To solve this, this commit simply adds a new foreach helper method that
avoids submitting corner elements to the provided callback.