Fix #128407: EEVEE: Cannot allocate volume pool on Intel ARC/OpenGL
When using OpenGL on Intel ARC the driver reports a max 3d allowed size
of 2048. The volume probe will create a texture that doesn't fit in this
dimension when selecting a probe size of 512 or 1024 MB.

This PR will change the atlas col_count to 512 so the row would fit
insize 2048. Adding support for 512MB pools. 1024MB pools can not be
supported in the current design.

According to https://opengl.gpuinfo.org/displaycapability.php?name=GL_MAX_3D_TEXTURE_SIZE
there are lower values but those are plaforms unable to run Blender (OpenGL 2.1)

The root cause is a part of the OpenGL spec which isn't respected by all
drivers:
```
GL_MAX_TEXTURE_SIZE, and GL_MAX_3D_TEXTURE_SIZE are not really
descriptive enough. It has to report the largest square texture image
that can be accommodated with mipmaps but a long skinny texture, or a
texture without mipmaps may easily fit in texture memory.
```
https://registry.khronos.org/OpenGL-Refpages/gl4/html/glGetTexLevelParameter.xhtml