Fix #128401: Issues when splitting connected strip
This patch fixes broken connections and effect strips which would arise
in some cases when trying to split connected strips with effects.

This happened in case an effect strip was being split -- first, its
effect chain would be added to a `VectorSet` of seqs, but then, only the
connected strips of the original seq (in this case the effect strip)
would be added to the `VectorSet`. Fix by iterating over the entire list
and adding connected strips.

Additionally:
* Add a final `SEQ_iterator_set_expand()` for `SEQ_query_effect_chain()`
  in case these connected strips had effects themselves.
* Move the `seq_edit_split_operation_permitted()` check after the entire
  chain is evaluated to avoid bugs w/ locked strips being editable
  before.