Fix: VSE timeline error w/ quick switch to boxsel
After removing box select in #128051, one keymap item remained that
would still attempt to switch to box select if the user tried to press W
in the timeline. Fix by adding two separate "set tool by name" keymap
items for sequencer and timeline, so that it sets to tweak if in the
timeline.

Also cleanup some old comments related to quick switch which are not
relevant anymore: when fallback tools were added in 6ffcddc as part of
2.82, left select could now easily select with any tool active.