Fix #128407: EEVEE: Cannot allocate volume pool on Intel ARC/OpenGL
When using OpenGL on Intel ARC the driver reports a max 3d allowed size
of 2048. The volume probe will create a texture that doesn't fit in this
dimension when selecting a probe size of 512 or 1024 MB.

This PR will allocate the max size allowed to allocate. Previous patch
added reshaping, but that introduces render artifacts. In order to fix
the crash it has been disabled for now.

Ref !128518: Previous implementation with reshaping enabled.