macOS/QuickLook: support rich thumbnail
Support thumbnail that shows the file contents instead
of the default blend file icon for all files in Finder/ Expose.
Some files may still have the usual blender icon as thumbnail
depending on their contents.

blender-thumbnailer process is kept alive by the system
in the background and is invoked by QuickLook when needed.

It will NOT be an interactive "Preview" that allows richer
interactions like Panning viewport, or rotating 3D objects.