Deps: build the `ffmpeg` commandline tool for Flamenco to harvest
As part of building Blender's dependencies, also build the FFmpeg CLI
tool. This is mostly to accommodate macOS/arm64, for which the FFmpeg
team do not provide precompiled binaries themselves. And I'm hesitatnt to
bundle a third-party ffmpeg binary with Flamenco.

Of course we don't build with all the fancy GPU encoding/decoding options,
but at least it's a better option than having no ffmpeg bundled at all.
Flamenco docs can describe how to replace it with a fancier one.