Fix #126541: Tooltip for dynamic custom enum properties broken
Basic issue here is that we're referencing strings from Python that have
their own lifetime. Python is not aware of this reference and will
garbage collect the strings (I guess).

There's a warning in the [`EnumProperty` Python API
documentation](https://docs.blender.org/api/current/bpy.props.html#bpy.props.EnumProperty)
about this. However, this case should be easy to fix by storing a copy
of the string instead of keeping a reference to it. `uiBut.tip` should
probably be converted into a `std::string` which would do just that, but
this is better to do in `main` than in the release branch.