Fix #128346: search results fail to display at first
A delay refreshing the outliner when changes content cause the outliner
to clamp the scroll position without showing the new scroll position.
The outliner would simply not show any results, making it seem as
though the search had no matches.

This is a regression since 2.7x which would calculate the current area
before drawing the outliner contents. Since then, the logic changed
to use limits calculated while drawing to set the total 2D region.

Failure to update also caused the the outliner to display as empty
in similar situations such as deleting objects in the 3D viewport.

Resolve by detecting when the scroll position changes and
drawing a second time.