Fix: ffmpeg VP9 lossless flag was wrongly applied to all WebM containers
Previous fix to make VP9 lossless work (98689f51c0114) applied it to
all videos that happen to be in WebM containers. While it is typical
that WebM would be used for VP9, it is not necessarily so (WebM can
have H.264 or really any other video). Do the check based on video
codec being VP9.